# server/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Cheese": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"colour": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"price",
					"colour",
					"description",
					"category",
					"image",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICheesePayload": {
				"properties": {
					"title": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"colour": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"price",
					"colour",
					"description",
					"category",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Purchase": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"cheeses": {
						"items": {
							"$ref": "#/components/schemas/Cheese"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userId",
					"totalPrice",
					"totalItems",
					"cheeses",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPurchasePayload": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"dateTime": {
						"type": "string"
					},
					"cheeses": {
						"items": {
							"$ref": "#/components/schemas/Cheese"
						},
						"type": "array"
					}
				},
				"required": [
					"userId",
					"totalPrice",
					"totalItems",
					"dateTime",
					"cheeses"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/cheeses": {
			"get": {
				"operationId": "GetCheeses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Cheese"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Cheese"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCheese",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Cheese"
								}
							}
						}
					}
				},
				"tags": [
					"Cheese"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICheesePayload"
							}
						}
					}
				}
			}
		},
		"/cheeses/{id}": {
			"get": {
				"operationId": "GetCheese",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Cheese"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cheese"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/cheeses/{id}/{weight}": {
			"get": {
				"operationId": "GetCheeseTotalPrice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double",
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cheese"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "weight",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/purchases/{userId}": {
			"get": {
				"operationId": "GetPurchases",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Purchase"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Purchase"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/purchases/create": {
			"post": {
				"operationId": "CreatePurchase",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Purchase"
								}
							}
						}
					}
				},
				"tags": [
					"Purchase"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPurchasePayload"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}